Fixed implicit narrowing conversion in compound assignments.

+++ src/com/digero/common/midi/TimeSignature.java
@@ -182,26 +182,26 @@ public class TimeSignature implements MidiConstants
 		if (n >= (1 << 16))
 		{
 			n >>>= 16;
-			pos += 16;
+			pos += (byte) 16;
 		}
 		if (n >= (1 << 8))
 		{
 			n >>>= 8;
-			pos += 8;
+			pos += (byte) 8;
 		}
 		if (n >= (1 << 4))
 		{
 			n >>>= 4;
-			pos += 4;
+			pos += (byte) 4;
 		}
 		if (n >= (1 << 2))
 		{
 			n >>>= 2;
-			pos += 2;
+			pos += (byte) 2;
 		}
 		if (n >= (1 << 1))
 		{
-			pos += 1;
+			pos += (byte) 1;
 		}
 		return ((n == 0) ? (-1) : pos);
 	}
